Add entropy estimation to the O0 histogram generation.
I'm not sure if this is worth it or not as the branchless version
generally works best of the most compressed rANS option, which is what
we'll end up automatically selecting.

It seems to have about 6% CPU hit for encoding O0 (due to the entropy
calc I assume), but otherwise it does a good job of rescuing the
substantial slow down for O0 encoding of very compressible data such
as 4-qual Illumina.