Improve compute_shift estimates.
The calculations are now done as doubles instead of going via int,
which was skewing the entropy calculation a bit.  Also measured the
frequency symbol table from some data to get real values.

The net effect on this changes appears to be to use 10-bit freqs more
often than before, so up to 1% growth for some marginal data series,
but considerably quicker on them too.