Add ARM64 Neon (SIMD) instructions for rANS32x16.
This speeds up decoding to be faster than the 4x16 codec, but encoding
is still a challenge.  With gcc on the q40 (40 illumina quals) test:

Gcc Order-0
4x scalar  345.9 MB/s enc, 446.0 MB/s dec	 101000000 -> 51136478 bytes
32x scalar 335.0 MB/s enc, 443.2 MB/s dec	 101000000 -> 51144602 bytes
32x neon   359.7 MB/s enc, 716.0 MB/s dec	 101000000 -> 51144602 bytes

Gcc Order-1
4x scalar  203.8 MB/s enc, 323.0 MB/s dec	 101000000 -> 48928863 bytes
32x scalar 161.1 MB/s enc, 256.5 MB/s dec	 101000000 -> 48937630 bytes
32x neon   178.2 MB/s enc, 397.6 MB/s dec	 101000000 -> 48937630 bytes

This shows the 32-way interleaving is slower than 4-way with scalar
code (poor auto-vectorisation, as expected), but manual vectorisation
is faster that both for decode, especially order-0.  However Neon
32-way order-1 encode is still slower than the old 4-way scalar,
making neon the 32-way unrolling questionable on ARM.

There is also a heavy gcc vs clang bias.

Clang Order-0

4x scalar  300.1 MB/s enc, 386.2 MB/s dec	 101000000 -> 51136478 bytes
32x scalar 291.8 MB/s enc, 421.1 MB/s dec	 101000000 -> 51144602 bytes
32x neon   296.6 MB/s enc, 525.0 MB/s dec	 101000000 -> 51144602 bytes

Here we see encode times all very similar to each other (as with gcc),
but Neon decode doesn't gain as much. Note the 32x scalar code is 5%
quicker on gcc than clang, but the neon version is 36% quicker on
gcc.  How can be it so different, given Neon intrinsics are basically
just assembly instructions?  The answer is that Clang-10 is reordering
the instructions suboptimally and introducing instruction latency.  I
cannot find a way to disable that (short of writing it as the
assembly).

Clang Order-1

4x scalar  196.9 MB/s enc, 345.5 MB/s dec	 101000000 -> 48928863 bytes
32x scalar 158.9 MB/s enc, 254.7 MB/s dec	 101000000 -> 48937630 bytes
32x neon   175.0 MB/s enc, 350.0 MB/s dec	 101000000 -> 48937630 bytes

Clang order-1 follows a similar pattern, with gcc giving a better
ordering of neon instructions, however the impact is less as order-1
is quite memory fetch bound. Again, order-1 32-way is a questionable
choice, even with Neon.