Add commented out attempts to SSE4 optimise the order-0 encoder.
This doesn't help (yet), but it's instructional maybe and it's left
there as a documented attempt.  For now the scalar 32-way encoder is
used when in an SSE4 environment.

This scalar version has been further optimised slightly, so the same
code works well now on both gcc and clang (and a bit better than
either version before).