Refactor the ARM Neon code a little bit.
This removes some of the common code, utils the static inlines instead
as per SSE4, AVX2 and AVX512 variants.  However the O1 decoder still
has the original code in place because for some reason gcc on ARM
generates slower code (by about 10%).  It's unknown why this happens
on ARM and not elsewhere.