Small optimisations to hist8 and hist1_4 functions.
On large buffers, the order-0 function now adds to 16-bit frequency
tables and at the end separates that out to 8-bit frequencies.

A similar technique was tried with order-1, but it wasn't universally
beneficial and on average poorer.  However we moved the T[x] frequency
to outside the main loop, which is generally a win.