Revamp the "order" parameter and test setup.
We no longer overload "order" to control the CPU methods.  This is now
a separate parameter and we've added a rans_set_cpu() function to
explicitly control this.

The order bit field macros are now also public, but haven't changed
bit values.

The N-stripe parameter is now once again bit-8 onwards rather than
being shifted in the unroll32 branch to bit-16.  This keeps
compatibility with master.