1) Added the ability to use sequence bases as a quality context.
The idea of using sequence as a context for quality value encoding
goes back a long way.  Certainly it was discussed during the
SequenceSqueeze contest:

https://encode.su/threads/1409-Compression-Competition-15-000-USD?p=28377&viewfull=1#post28377

Ultimately it didn't help much in fqzcomp because sequence is not a
major player in Illumina quality value compression.  The sequence (in
a wider context than just 1 bp) is however highly indicative of
quality in PacBio CLR and ONT data.

An example of adding sequence as a context to some PB data:

    ./tests/fqzcomp_qual -x 0x660000000000000 ~/tmp/PB.qual
    Total output = 14266164

    ./tests/fqzcomp_qual -x 0x660000000000862 ~/tmp/PB.qual ~/tmp/PB.seq
    Total output = 11903124

Unfortunately this isn't available within CRAM yet as it means added
correlations between the streams.  Although it's curious this saves
2363040 bytes on 26885686 values, so 0.7 bit per value.  Still
considerably lower than the size of compressed sequence (about
1.7bit/base for this example), but there may even be occasions where
storing a squashed sequence hash in the stream can improve compression.

TODO: figure out how to add to CRAM 3.1 / 4.0.
- We need to cope with reversing quality.  This also needs revcomp on seq.
- How to detect when it's useful.  Long variable read length as a proxy?

2) Add FQZcomp GFLAG_USE_SEQ.

This is an easy way to detect that sequence contexts are in use, to
aid CRAM decoding. The alternative would be a different FQZ name, eg
FQS, so we do it by codec variant.  However the current solution is
more in line with O0/O1 for rANS, albeit that's not something the
decoder needs to be aware of.

3) Tweaks to standard fqzcomp parameter sets

- Add qbits usage for PacBio CLR.

The huge number of quality symbols makes this a considerable
improvement.

Before at fqzcomp5 -Q3: 774038602 to 196795992
After, with qmap:       774038602 to 181705946

- Add an Ultimate Genomics specific fqzcomp profile

This makes use of the new ability to utilise sequence context.