Add FQZcomp GFLAG_USE_SEQ.
This is an easy way to detect that sequence contexts are in use, to
aid CRAM decoding. The alternative would be a different FQZ name, eg
FQS, so we do it by codec variant.  However the current solution is
more in line with O0/O1 for rANS, albeit that's not something the
decoder needs to be aware of.