Substantial reduction to name tokeniser memory.
We allocate last_context based on the current number of tokens used
rather than MAX_TOKENS, although temporarily it is at this size and
realloced down at the end.

The trie_t structure has been rejigged to be 24 bytes instead of 32
bytes.  This is possible as MAX_NAMES is ~20 bits so we can use
bit-fields.