Bug fix the -c <DEC><ENC> option for testing codecs.
The selection of encode vs decode codecs to use is controlled by the
CPU option -c which has two bytes for setting the CPU type to control
the decoder and encoder.  This was changed in 958032c to only do CPU
detection once.

This commit moved the have_avx2 etc variables out from rans_enc_func
and rans_dec_func to be globals, initialised once.

However the original code had

    if (!(rans_cpu & RANS_CPU_ENC_AVX512)) have_avx512f = 0;

and

    if (!(rans_cpu & RANS_CPU_DEC_AVX512)) have_avx512f = 0;

Note the difference between _ENC_ and _DEC_, giving us the ability to
configure each side independently.  We use this in the tests/entropy
tool for cross-validation to ensure codecs are compatible with one
another.  When made global, this distinction was lost and only the
encoder side was used, meaning we lost that cross-validation.

This has been resurrected by moving the rans_cpu filtering back into
the rans_dec_func and rans_enc_func functions.

+debugging. DEL