Add a tok3 and fqzcomp round-trip fuzzer
We may also need to explore choice of codec more so we don't always
choose the smallest, as with tiny fuzzing sizes this limits the range
of codecs being tested in the decompression / round-trip stage.

Also make fqzcomp encoder better cope with malformed input paramsters,
where we s->num_records and/or length summation doesn't match the
input buffer.