Replace _mm512_i32gather_epi32(mm512, array, 1) with simulated gather.
This also fixes a bug where we read beyond an array by up to 3 bytes
and then AND off the top 3-bytes again.  The old epi32 gather loads
32-bit quantities, at 1-byte size offsets.  We then have to AND off
all bar the bottom 8 bits as we wanted 8-bit quantities.  For the
final part of the "in" buffer this means we may overflow the memory by
up to 3 bytes, causing a rare SEGV.